refactor : 기존 error 패키지 하위 클래스를 exception으로 이동
