+++ src/main/java/com/clubber/ClubberServer/domain/admin/exception/AdminAlreadyDeletedException.java
@@ -1,6 +1,6 @@
 package com.clubber.ClubberServer.domain.admin.exception;
 
-import com.clubber.ClubberServer.global.error.BaseException;
+import com.clubber.ClubberServer.global.exception.BaseException;
 
 public class AdminAlreadyDeletedException extends BaseException {
