+++ src/main/java/com/clubber/ClubberServer/domain/admin/exception/AdminEqualsPreviousPasswordExcpetion.java
@@ -1,6 +1,6 @@
 package com.clubber.ClubberServer.domain.admin.exception;
 
-import com.clubber.ClubberServer.global.error.BaseException;
+import com.clubber.ClubberServer.global.exception.BaseException;
 
 public class AdminEqualsPreviousPasswordExcpetion extends BaseException {
 	public static final BaseException EXCEPTION = new AdminEqualsPreviousPasswordExcpetion();
