+++ src/main/java/com/clubber/ClubberServer/domain/admin/exception/AdminNotFoundException.java
@@ -1,7 +1,7 @@
 package com.clubber.ClubberServer.domain.admin.exception;
 
 
-import com.clubber.ClubberServer.global.error.BaseException;
+import com.clubber.ClubberServer.global.exception.BaseException;
 
 public class AdminNotFoundException extends BaseException {
