+++ src/main/java/com/clubber/ClubberServer/domain/club/exception/ClubAlreadyDeletedException.java
@@ -1,7 +1,7 @@
 package com.clubber.ClubberServer.domain.club.exception;
 
 
-import com.clubber.ClubberServer.global.error.BaseException;
+import com.clubber.ClubberServer.global.exception.BaseException;
 
 public class ClubAlreadyDeletedException extends BaseException {
