+++ src/main/java/com/clubber/ClubberServer/domain/club/exception/ClubNotFoundException.java
@@ -1,8 +1,6 @@
 package com.clubber.ClubberServer.domain.club.exception;
 
-import com.clubber.ClubberServer.domain.favorite.exception.ClubAlreadyRegisterdFavoriteException;
-import com.clubber.ClubberServer.domain.favorite.exception.FavoriteErrorCode;
-import com.clubber.ClubberServer.global.error.BaseException;
+import com.clubber.ClubberServer.global.exception.BaseException;
 
 public class ClubNotFoundException extends BaseException{
     public static final BaseException EXCEPTION = new ClubNotFoundException();
