+++ src/main/java/com/clubber/ClubberServer/domain/club/exception/DepartmentNotFoundException.java
@@ -1,6 +1,6 @@
 package com.clubber.ClubberServer.domain.club.exception;
 
-import com.clubber.ClubberServer.global.error.BaseException;
+import com.clubber.ClubberServer.global.exception.BaseException;
 
 public class DepartmentNotFoundException extends BaseException {
     public static final BaseException EXCEPTION = new DepartmentNotFoundException();
