+++ src/main/java/com/clubber/ClubberServer/domain/club/exception/DivisionNotFoundException.java
@@ -1,6 +1,6 @@
 package com.clubber.ClubberServer.domain.club.exception;
 
-import com.clubber.ClubberServer.global.error.BaseException;
+import com.clubber.ClubberServer.global.exception.BaseException;
 
 public class DivisionNotFoundException extends BaseException {
     public static final BaseException EXCEPTION = new DivisionNotFoundException();
