+++ src/main/java/com/clubber/ClubberServer/domain/favorite/exception/FavoriteNotMatchClubException.java
@@ -1,7 +1,6 @@
 package com.clubber.ClubberServer.domain.favorite.exception;
 
-import com.clubber.ClubberServer.global.error.BaseErrorCode;
-import com.clubber.ClubberServer.global.error.BaseException;
+import com.clubber.ClubberServer.global.exception.BaseException;
 
 public class FavoriteNotMatchClubException extends BaseException {
