+++ src/main/java/com/clubber/ClubberServer/domain/notice/exception/NoticeNotFoundException.java
@@ -1,8 +1,6 @@
 package com.clubber.ClubberServer.domain.notice.exception;
 
-import com.clubber.ClubberServer.domain.club.exception.ClubErrorCode;
-import com.clubber.ClubberServer.domain.club.exception.ClubIdNotFoundException;
-import com.clubber.ClubberServer.global.error.BaseException;
+import com.clubber.ClubberServer.global.exception.BaseException;
 
 public class NoticeNotFoundException extends BaseException {
     public static final BaseException EXCEPTION = new NoticeNotFoundException();
