+++ src/main/java/com/clubber/ClubberServer/domain/review/exception/ReviewClubNotMatchException.java
@@ -1,7 +1,6 @@
 package com.clubber.ClubberServer.domain.review.exception;
 
-import com.clubber.ClubberServer.global.error.BaseErrorCode;
-import com.clubber.ClubberServer.global.error.BaseException;
+import com.clubber.ClubberServer.global.exception.BaseException;
 
 public class ReviewClubNotMatchException extends BaseException {
