+++ src/main/java/com/clubber/ClubberServer/domain/review/exception/ReviewErrorCode.java
@@ -1,7 +1,7 @@
 package com.clubber.ClubberServer.domain.review.exception;
 
-import com.clubber.ClubberServer.global.error.BaseErrorCode;
-import com.clubber.ClubberServer.global.error.ErrorReason;
+import com.clubber.ClubberServer.global.exception.BaseErrorCode;
+import com.clubber.ClubberServer.global.exception.ErrorReason;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
