+++ src/main/java/com/clubber/ClubberServer/domain/review/exception/UserAlreadyReviewedException.java
@@ -1,6 +1,6 @@
 package com.clubber.ClubberServer.domain.review.exception;
 
-import com.clubber.ClubberServer.global.error.BaseException;
+import com.clubber.ClubberServer.global.exception.BaseException;
 
 public class UserAlreadyReviewedException extends BaseException {
