+++ src/main/java/com/clubber/ClubberServer/domain/review/exception/UserReviewsNotFoundException.java
@@ -1,6 +1,6 @@
 package com.clubber.ClubberServer.domain.review.exception;
 
-import com.clubber.ClubberServer.global.error.BaseException;
+import com.clubber.ClubberServer.global.exception.BaseException;
 
 public class UserReviewsNotFoundException extends BaseException{
     public static final BaseException EXCEPTION = new UserReviewsNotFoundException();
