+++ src/main/java/com/clubber/ClubberServer/domain/user/exception/RefreshTokenExpiredException.java
@@ -1,6 +1,6 @@
 package com.clubber.ClubberServer.domain.user.exception;
 
-import com.clubber.ClubberServer.global.error.BaseException;
+import com.clubber.ClubberServer.global.exception.BaseException;
 
 public class RefreshTokenExpiredException extends BaseException {
