+++ src/main/java/com/clubber/ClubberServer/global/config/response/GlobalExceptionHandler.java
@@ -2,10 +2,10 @@ package com.clubber.ClubberServer.global.config.response;
 
 import java.util.List;
 
-import com.clubber.ClubberServer.global.error.BaseErrorCode;
-import com.clubber.ClubberServer.global.error.BaseException;
-import com.clubber.ClubberServer.global.error.ErrorReason;
-import com.clubber.ClubberServer.global.error.GlobalErrorCode;
+import com.clubber.ClubberServer.global.exception.BaseErrorCode;
+import com.clubber.ClubberServer.global.exception.BaseException;
+import com.clubber.ClubberServer.global.exception.ErrorReason;
+import com.clubber.ClubberServer.global.exception.GlobalErrorCode;
 import com.clubber.ClubberServer.global.event.exceptionalarm.ExceptionAlarmEventPublisher;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
