+++ src/main/java/com/clubber/ClubberServer/global/exception/EnumTypeNotValidException.java
@@ -1,8 +1,5 @@
 package com.clubber.ClubberServer.global.exception;
 
-import com.clubber.ClubberServer.global.error.BaseException;
-import com.clubber.ClubberServer.global.error.GlobalErrorCode;
-
 public class EnumTypeNotValidException extends BaseException {
 
 	public static final BaseException EXCEPTION = new EnumTypeNotValidException();
